Moved the location of the student and faculty pages to be in their own packages within pages.
