+++ john_callahan_p0/src/main/java/com/revature/projectZero/pages/student/StudentRegisterPage.java
@@ -1,5 +1,6 @@
-package com.revature.projectZero.pages;
+package com.revature.projectZero.pages.student;
 
+import com.revature.projectZero.pages.Page;
 import com.revature.projectZero.pojos.Student;
 import com.revature.projectZero.service.ValidationService;
 import com.revature.projectZero.util.PageRouter;
