+++ john_callahan_p0/src/main/java/com/revature/projectZero/util/AppState.java
@@ -3,6 +3,9 @@ package com.revature.projectZero.util;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import com.revature.projectZero.pages.*;
+import com.revature.projectZero.pages.faculty.FacultyDashboard;
+import com.revature.projectZero.pages.student.StudentDashboard;
+import com.revature.projectZero.pages.student.StudentRegisterPage;
 import com.revature.projectZero.repositories.SchoolRepository;
 import com.revature.projectZero.service.ValidationService;
 import org.apache.logging.log4j.LogManager;
